Update README.md
I've made the recommended changes and added the $$$ to our entry.  I request, however, that the same notation be added to the other vendors who all have paid offerings, including New Relic, Dynatrace, and AppDynamics.  I notice that none of them have that notation.  Also, the Dynatrace entry is in the original section I chose (Monitoring & Logging) and should also be moved to Monitoring & Logging Services as they are a similar service to CA, AppD, and New Relic.

Thanks for building this page!